+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - `add_minor_chord()`
 - demos module:
     - `demo1()`
+- envelopes module: `ADSR_envelope()` function
 
 ## [Forsynth 0.1 "Stockhausen"] 2021-02-19